+++ composer.json
@@ -1,5 +1,6 @@
 {
   "name": "digitalpolygon/polymer-drupal",
+  "type": "polymer-plugin",
   "description": "Drupal integration for Polymer",
   "license": "GPL-2.0+",
   "authors":[
@@ -29,7 +30,7 @@
     }
   },
   "require": {
-    "digitalpolygon/polymer": "0.x-dev",
+    "digitalpolygon/polymer": "dev-feature/refactor-organization",
     "digitalpolygon/drupal-upgrade-plugin": "^2@dev",
     "drush/drush": "^12 || ^13",
     "php": ">=8.1",
@@ -43,13 +44,14 @@
   },
   "minimum-stability": "dev",
   "prefer-stable": true,
-  "autoload": {
-    "psr-4": {
-      "DigitalPolygon\\PolymerDrupal\\": "./src/"
-    }
+  "extra": {
+    "installer-name": "polymer_drupal"
   },
   "config": {
     "allow-plugins": {
+      "composer/installers": true,
+      "mnsami/composer-custom-directory-installer": true,
+      "oomphinc/composer-installers-extender": true,
       "digitalpolygon/polymer": true,
       "digitalpolygon/drupal-upgrade-plugin": true,
       "phpro/grumphp-shim": true
@@ -63,6 +65,11 @@
       "@lint",
       "@cs",
       "@sa"
+    ]
+  },
+  "autoload": {
+    "classmap": [
+      "src/VendorAssets.php"
     ]
   }
 }
