+++ phpstanBootstrap.php
@@ -1,0 +1,3 @@
+<?php
+
+define('DRUPAL_ROOT', '/var/www/html/web');
