+++ settings/polymer.settings.php
@@ -1,13 +1,13 @@
 <?php
 
-use DigitalPolygon\Polymer\Environment\EnvironmentDetectorBase;
-use DigitalPolygon\Polymer\Environment\AcquiaEnvironmentDetector;
-use DigitalPolygon\Polymer\Environment\PantheonEnvironmentDetector;
-
 /**
  * @file
  * Generated by Polymer.
  */
+
+use DigitalPolygon\Polymer\Core\Environment\EnvironmentDetectorBase;
+use DigitalPolygon\Polymer\Core\Environment\AcquiaEnvironmentDetector;
+use DigitalPolygon\Polymer\Core\Environment\PantheonEnvironmentDetector;
 
 /**
  * Site path.
@@ -17,12 +17,6 @@ use DigitalPolygon\Polymer\Environment\PantheonEnvironmentDetector;
  */
 $site_name = str_replace('sites/', '', $site_path);
 
-/**
- * Site path.
- *
- * @var string $repo_root
- *  Path to the repository root (assumed to be parent of DRUPAL_ROOT).
- */
 if (!isset($repo_root)) {
     $repo_root = realpath(DRUPAL_ROOT . '/..');
 }
@@ -35,12 +29,10 @@ $polymer_settings_files = [
     'config',
     'db',
 ];
-if ($polymer_settings_files) {
-    foreach ($polymer_settings_files as $polymer_settings_file) {
-        $settings_file = __DIR__ . "/$polymer_settings_file.settings.php";
-        if (file_exists($settings_file)) {
-            $settings_files[] = $settings_file;
-        }
+foreach ($polymer_settings_files as $polymer_settings_file) {
+    $settings_file = __DIR__ . "/$polymer_settings_file.settings.php";
+    if (file_exists($settings_file)) {
+        $settings_files[] = $settings_file;
     }
 }
